根据计算表达式得到各个子模块的结果，避免了硬编码问题，目前投资明细表计算结果没问题，暂时未解决一个子模块（如综合评估表的投入合计OT034）包含多个计算表达式的情况。

+++ snowy-plugin/snowy-plugin-biz/src/main/java/vip/xiaonuo/biz/modular/New/subtemplate/dto/SubtemplateInputParam.java
@@ -15,8 +15,8 @@ public class SubtemplateInputParam {
     @NoArgsConstructor
     @Data
     public static class SubmouleDV {
-        private Integer subtemplateId;
-        private Integer projectId;
+        private Long subtemplateId;
+        private Long projectId;
         private List<Data> data;
 
         @NoArgsConstructor
